requirements.txt
py-clob-client==0.34.6
web3==6.14.0
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.7
schedule==1.2.1
structlog==24.1.0
numpy==1.26.4
